refactor: move auth into feature module
