+++ frontend/src/app/router.tsx
@@ -1,7 +1,6 @@
 import { Navigate, createBrowserRouter } from "react-router-dom";
 import CanvasPage from "@/features/gameplay/pages/CanvasPage";
-import LoginPage from "@/pages/LoginPage";
-import RegisterPage from "@/pages/RegisterPage";
+import { LoginPage, RegisterPage } from "@/features/auth";
 
 export const router = createBrowserRouter([
   {
