+++ frontend/src/features/auth/index.ts
@@ -1,0 +1,6 @@
+export { authApi } from "./api/auth.api";
+
+export type { RegisterRequest, LoginRequest, Voter } from "./api/auth.api";
+
+export { default as LoginPage } from "./pages/LoginPage";
+export { default as RegisterPage } from "./pages/RegisterPage";
