+++ frontend/src/features/auth/pages/LoginPage.tsx
@@ -1,13 +1,12 @@
 import { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
-import { authApi } from "@/api/auth";
 import { Button } from "@/shared/ui/button";
+import { authApi } from "../api/auth.api";
 
-export default function RegisterPage() {
+export default function LoginPage() {
   const navigate = useNavigate();
   const [username, setUsername] = useState("");
   const [password, setPassword] = useState("");
-  const [nickname, setNickname] = useState("");
   const [error, setError] = useState("");
 
   const handleSubmit = async (e: React.FormEvent) => {
@@ -15,14 +14,14 @@ export default function RegisterPage() {
     setError("");
 
     try {
-      await authApi.register({ username, password, nickname });
-      navigate("/login");
+      await authApi.login({ username, password });
+      navigate("/canvas");
     } catch (err: unknown) {
       if (err && typeof err === "object" && "response" in err) {
         const axiosErr = err as { response: { data: { message: string } } };
         setError(axiosErr.response.data.message);
       } else {
-        setError("회원가입 중 오류가 발생했어요.");
+        setError("로그인 중 오류가 발생했어요.");
       }
     }
   };
@@ -45,17 +44,6 @@ export default function RegisterPage() {
           </div>
 
           <div className="flex flex-col gap-1">
-            <label className="text-sm font-medium">닉네임</label>
-            <input
-              type="text"
-              value={nickname}
-              onChange={(e) => setNickname(e.target.value)}
-              className="rounded border px-3 py-2 text-sm"
-              required
-            />
-          </div>
-
-          <div className="flex flex-col gap-1">
             <label className="text-sm font-medium">비밀번호</label>
             <input
               type="password"
@@ -69,14 +57,14 @@ export default function RegisterPage() {
           {error && <p className="text-sm text-red-500">{error}</p>}
 
           <Button type="submit" className="w-full">
-            회원가입
+            로그인
           </Button>
         </form>
 
         <p className="text-center text-sm">
-          이미 계정이 있으신가요?{" "}
-          <Link to="/login" className="underline">
-            로그인
+          계정이 없으신가요?{" "}
+          <Link to="/register" className="underline">
+            회원가입
           </Link>
         </p>
       </div>
